feat(#1146): Split explorable maps and boss domain maps drop tables
