+++ Content/Items/Consumables/Maps/ExplorableMaps/ExplorableMap.cs
@@ -1,0 +1,6 @@
+﻿namespace PathOfTerraria.Content.Items.Consumables.Maps.ExplorableMaps;
+
+public abstract class ExplorableMap : Map
+{
+	
+}
